+++ src/charts/tests.rs
@@ -97,6 +97,7 @@ fn plot_metrics_creates_files() -> Result<(), String> {
             method: HttpMethod::Get,
             url: Some("http://localhost".to_owned()),
             headers: vec![],
+            no_ua: false,
             data: String::new(),
             target_duration: PositiveU64::try_from(1)?,
             expected_status_code: 200,