+++ src/config/types.rs
@@ -18,6 +18,7 @@ pub struct ConfigFile {
     pub status: Option<u16>,
     pub charts_path: Option<String>,
     pub no_charts: Option<bool>,
+    pub no_ua: Option<bool>,
     pub tmp_path: Option<String>,
     pub keep_tmp: Option<bool>,
     pub export_csv: Option<String>,