+++ src/distributed/tests.rs
@@ -21,6 +21,7 @@ fn base_args(url: String, tmp_path: String) -> Result<TesterArgs, String> {
         method: HttpMethod::Get,
         url: Some(url),
         headers: vec![],
+        no_ua: false,
         data: String::new(),
         target_duration: positive_u64(1)?,
         expected_status_code: 200,